Fixed typing (linting) for timings to allow float values.

+++ src/jobflow_remote/jobs/jobcontroller.py
@@ -4393,7 +4393,7 @@ class JobController:
         self,
         query: dict,
         max_step_attempts: int,
-        delta_retry: tuple[int, ...],
+        delta_retry: tuple[float, ...],
         next_step_delay: int | None = None,
         **kwargs,
     ) -> Generator[MongoLock, None, None]:
